# DOT-LMS-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOT-LMS API Documentation",
        "contact": {
            "email": "ruthgetaneh5@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "DOT-LMS API Server"
        }
    ],
    "paths": {
...
        "/Login": {
            "post": {
                "tags": [
                    "User Management Endpoints"
                ],
                "summary": "Login a user into the system",
                "description": "Login a user into the system",
                "operationId": "5426fe18441cdca79e9e3c58873c3582",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "password"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "ADM-7864"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/ConfigurePassword": {
            "post": {
                "tags": [
                    "User Management Endpoints"
                ],
                "summary": "configure a user password",
                "description": "an API Endpoint that enables a user to configure a password, otp=> one time passcode",
                "operationId": "3db665430fd0cfdb182c83d81aded7d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "otp",
                                    "password"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "ADM-7864"
                                    },
                                    "otp": {
                                        "type": "integer",
                                        "example": "234123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/forgotPassword": {
            "post": {
                "tags": [
                    "User Management Endpoints"
                ],
                "summary": "user enters his/her ID to get a configure password email",
                "description": "an API Endpoint that enables a user to configure a password,if they forgot their password, otp=> one time passcode",
                "operationId": "e83bc8ca0ff0085b50fbfc5bbb0742df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "ADM-7864"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "api_key",
                "in": "header"
            }
        }
    },
...
}